Se refactoriza codigo de fase1, se agregan interfaces, se mejora inyeccion de dependencia y se agregan comentarios en metodos
