+++ TemplateTPCorto/Negocio/LoginNegocio.cs
@@ -1,6 +1,8 @@
 ﻿using Datos;
+using Negocio.interfaces;
 using Persistencia;
 using Persistencia.DataBase;
+using Persistencia.interfaces;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
@@ -12,16 +14,15 @@ using System.Threading.Tasks;
 
 namespace Negocio
 {
-    public class LoginNegocio
+    public class LoginNegocio:ILoginNegocio
     {
-        private readonly UsuarioPersistencia usuarioPersistencia;
+        private readonly IUsuarioPersistencia usuarioPersistencia;
         private const int MAX_INTENTOS = 3;
 
-        public LoginNegocio(UsuarioPersistencia usuarioPersistencia)
+        public LoginNegocio(IUsuarioPersistencia usuarioPersistencia)
         {
             this.usuarioPersistencia = usuarioPersistencia;
         }
-
         public Credencial Login(String usuario, String password)
         {
             Credencial credencial = usuarioPersistencia.Login(usuario);
@@ -37,10 +38,6 @@ namespace Negocio
             }
             return null;
         }
-        public Perfil ObtenerPerfil(string legajo)
-        {
-            return usuarioPersistencia.ObtenerPerfil(legajo);
-        }
         public bool EstaBloqueado(string usuario)
         {
             Credencial credencial = usuarioPersistencia.Login(usuario);
@@ -51,7 +48,15 @@ namespace Negocio
             }
             return false;
         }
-
+        public bool EsContraseniaExpirada(Credencial credencial)
+        {
+            return credencial.FechaUltimoLogin.HasValue &&
+                 credencial.FechaUltimoLogin.Value.AddDays(30) <= DateTime.Today;
+        }
+        public bool EsPrimerLogin(Credencial credencial)
+        {
+            return credencial.FechaUltimoLogin == default(DateTime);
+        }
         private void RegistrarIntento(Credencial credencial)
         {
             int intentos = usuarioPersistencia.ObtenerNumeroIntentosPorLegajo(credencial.Legajo);
@@ -65,150 +70,5 @@ namespace Negocio
                 usuarioPersistencia.BloquearUsuario(credencial.Legajo);
             }
         }
-
-        public bool EsContraseniaExpirada(Credencial credencial)
-        {
-            return credencial.FechaUltimoLogin.HasValue &&
-                 credencial.FechaUltimoLogin.Value.AddDays(30) <= DateTime.Today;
-        }
-
-        public bool EsPrimerLogin(Credencial credencial)
-        {
-            return credencial.FechaUltimoLogin == default(DateTime);
-        }
-        
-        public void ActualizarContrasenia(Credencial credencial)
-        {
-            usuarioPersistencia.ActualizarContrasenia(credencial);
-        }
-
-        public Credencial BuscarCredencialPorNumeroLegajo(String legajo)
-        {
-            return usuarioPersistencia.BuscarCredencialPorNumeroLegajo(legajo);
-        }
-
-        public Persona BuscarPersonaPorNumeroLegajo(string legajo)
-        {
-            return usuarioPersistencia.BuscarPersonaPorNumeroLegajo(legajo);
-        }
-
-        public void RegistrarOperacionCambioCredencial(Autorizacion autorizacion, OperacionCambioCredencial operacion)
-        {
-            string idOperacion = usuarioPersistencia.CrearAutorizacion(autorizacion);
-            operacion.IdOperacion = idOperacion;
-            usuarioPersistencia.RegistrarOperacionCambioCredencial(operacion);
-        }
-        public void RegistrarOperacionCambioPersona(Autorizacion autorizacion, OperacionCambioPersona operacion)
-        {
-            string idOperacion = usuarioPersistencia.CrearAutorizacion(autorizacion);
-            operacion.IdOperacion = idOperacion;
-            usuarioPersistencia.RegistrarOperacionCambioPersona(operacion);
-        }
-
-        public List<OperacionCambioCredencial> ObtenerOperacionesCambioCredencialPendientesAutorizar()
-        {
-            List<Autorizacion> autorizaciones = ObtenerAutorizacionesPorTipoOperacionEstado(
-                EnumTipoOperacion.CambioCredencial.ToString(),
-                EnumEstadoAutorizacion.Pendiente.ToString()
-            );
-            List<String> idsOperacion = ObtenerIdsOperacion(autorizaciones);
-            return usuarioPersistencia.ObtenerOperacionesCambioCredencialPorIdsOperacion(idsOperacion);
-        }
-
-        public List<OperacionCambioPersona> ObtenerOperacionesCambioPersonaPendientesAutorizar()
-        {
-            List<Autorizacion> autorizaciones = ObtenerAutorizacionesPorTipoOperacionEstado(
-                EnumTipoOperacion.CambioPersona.ToString(),
-                EnumEstadoAutorizacion.Pendiente.ToString()
-            );
-            List<String> idsOperacion = ObtenerIdsOperacion(autorizaciones);
-            return usuarioPersistencia.ObtenerOperacionesCambioPersonaPorIdsOperacion(idsOperacion);
-        }
-        
-        public void AutorizarOperacionesCambioCredencial(List<OperacionCambioCredencial> operaciones, string legajoAutorizador)
-        {
-            foreach (OperacionCambioCredencial row in operaciones)
-            {
-                Credencial credencial = usuarioPersistencia.ObtenerCredencialPorLegajo(row.Legajo);
-                credencial.Contrasena = row.Contrasena;
-                credencial.FechaUltimoLogin = null;
-                ActualizarContrasenia(credencial);
-                usuarioPersistencia.ReiniciarIntentos(row.Legajo);
-                usuarioPersistencia.EliminarUsuarioBloqueadoPorLegajo(row.Legajo);
-                ActualizarEstadoAutorizacion(
-                    row.IdOperacion,
-                    EnumEstadoAutorizacion.Autorizado.ToString(),
-                    legajoAutorizador
-                );
-            }
-        }
-        
-        public void AutorizarOperacionesCambioPersona(List<OperacionCambioPersona> operaciones, string legajoAutorizador)
-        {
-            foreach (OperacionCambioPersona row in operaciones)
-            {
-                Persona modificada = new Persona
-                {
-                    Legajo = row.Legajo,
-                    Nombre = row.Nombre,
-                    Apellido = row.Apellido,
-                    Dni = row.Dni,
-                    FechaIngreso = row.FechaIngreso
-                };
-                usuarioPersistencia.ModificarPersonaPorLegajo(modificada);
-                ActualizarEstadoAutorizacion(
-                    row.IdOperacion,
-                    EnumEstadoAutorizacion.Autorizado.ToString(),
-                    legajoAutorizador
-                );
-            }
-        }
-
-        public void RechazarOperacionesCambioCredencial(List<OperacionCambioCredencial> operaciones, string legajoAutorizador)
-        {
-            foreach (OperacionCambioCredencial row in operaciones)
-            {
-                ActualizarEstadoAutorizacion(
-                    row.IdOperacion,
-                    EnumEstadoAutorizacion.Rechazado.ToString(),
-                    legajoAutorizador
-                );
-            }
-        }
-
-        public void RechazarOperacionesCambioPersona(List<OperacionCambioPersona> operaciones, string legajoAutorizador)
-        {
-            foreach (OperacionCambioPersona row in operaciones)
-            {
-                ActualizarEstadoAutorizacion(
-                    row.IdOperacion,
-                    EnumEstadoAutorizacion.Rechazado.ToString(),
-                    legajoAutorizador
-                );
-            }
-        }
-        private List<Autorizacion> ObtenerAutorizacionesPorTipoOperacionEstado(string tipoOperacion, string estado)
-        {
-            return usuarioPersistencia.ObtenerAutorizacionesPorTipoOperacionEstado(tipoOperacion, estado);
-        }
-
-        private List<String> ObtenerIdsOperacion(List<Autorizacion> autorizaciones)
-        {
-            List<String> idsOperacion = new List<String>();
-            foreach (Autorizacion registro in autorizaciones)
-            {
-                idsOperacion.Add(registro.IdOperacion);
-            }
-            return idsOperacion;
-        }
-        private void ActualizarEstadoAutorizacion(string idOperacion, string estado, string legajoAutorizador)
-        {
-            Autorizacion autorizacion = usuarioPersistencia.ObtenerAutorizacionPorIdOperacion(idOperacion);
-            autorizacion.Estado = estado;
-            autorizacion.LegajoAutorizador = legajoAutorizador;
-            autorizacion.FechaAutorizacion = DateTime.Now;
-            usuarioPersistencia.ActualizarEstadoAutorizacion(autorizacion);
-        }
-
     }
 }
