+++ TemplateTPCorto/TemplateTPCorto/TemplateTPCorto.csproj
@@ -46,6 +46,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="FabricaFormularios.cs" />
     <Compile Include="FormAutorizaciones.cs">
       <SubType>UserControl</SubType>
     </Compile>
